+++ src/Generators/FakeIdGenerator.php
@@ -24,8 +24,8 @@ class FakeIdGenerator {
 		// Generates a random string of length 1-2.
 		$id = base_convert( (string)rand( 0, 36 ^ 2 ), 10, 36 );
 
-		// This will "compress" the uniqid (some sort of microtimestamp) to a more dense string
-		$id .= base_convert( uniqid(), 10, 36 );
+		// This will "compress" the uniqid (a hex string) to a more dense string
+		$id .= base_convert( uniqid(), 16, 36 );
 
 		return $id;
 	}